session

+++ PHPCRUD/updatedata.php
@@ -1,4 +1,15 @@
 <?php
+session_start();
+
+// Проверяем, существует ли сессия с именем пользователя
+if (!isset($_SESSION['username'])) {
+    // Если сессия с именем пользователя не существует, перенаправляем пользователя на страницу входа
+    header('Location: php/login.php');
+    exit(); // После перенаправления следует завершить выполнение скрипта
+}
+?>
+
+<?php
     require_once "conn.php";
     $id = $_GET["id"];
     $query = "DELETE FROM results WHERE id = '$id'";
